simplify by removing unrequired component wrapping
